+++ modules/framework/galasa-parent/dev.galasa.framework/src/main/java/dev/galasa/framework/internal/ras/FrameworkMultipleResultArchiveStore.java
@@ -142,7 +142,7 @@ public class FrameworkMultipleResultArchiveStore implements IResultArchiveStoreS
     }
 
     @Override
-    public int retrieveRunLogLineCount() {
+    public long retrieveRunLogLineCount() {
         throw new UnsupportedOperationException("Unimplemented method 'retrieveRunLogLineCount'");
     }
 }
