+++ modules/framework/galasa-parent/dev.galasa.framework/src/testFixtures/java/dev/galasa/framework/mocks/MockIResultArchiveStore.java
@@ -98,7 +98,7 @@ public class MockIResultArchiveStore implements IResultArchiveStore {
     }
 
     @Override
-    public int retrieveRunLogLineCount() {
+    public long retrieveRunLogLineCount() {
                 throw new UnsupportedOperationException("Unimplemented method 'retrieveRunLogLineCount'");
     }
 }
